// src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      high_roller_bets: {
        Row: {
          amount: number
          bet_data: Json
          bet_id: string
          created_at: string
          currency: string
          id: string
          status: string | null
        }
        Insert: {
          amount: number
          bet_data: Json
          bet_id: string
          created_at?: string
          currency: string
          id?: string
          status?: string | null
        }
        Update: {
          amount?: number
          bet_data?: Json
          bet_id?: string
          created_at?: string
          currency?: string
          id?: string
          status?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}